requirements.txt
djangorestframework==3.12.2
Django===4.0.1
asgiref===3.4.1
pytz==2020.5
